refactor(api): extract shared zip helpers into zip-util
Move createZipStream / collectFilePaths out of export-service into a shared
zip-util module and add two variants the part feature needs: in-memory extra
entries (for part.json + a rewritten config.yaml) and createZipStreamFromEntries
for a fixed entry list (the lightweight part archive). No behavior change for
the existing project/adt/webpub/epub/scorm exports.
